refactored using routers
