+++ TodoApp/routers/todos.py
@@ -1,0 +1,95 @@
+from typing import Annotated
+
+from pydantic import BaseModel, Field
+from starlette import status
+
+from models import Todos
+from database import SessionLocal
+
+from fastapi import Depends, HTTPException, Path, APIRouter
+from sqlalchemy.orm import Session
+
+router = APIRouter()
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+
+db_dependency = Annotated[Session, Depends(get_db)]
+
+
+class TodoRequest(BaseModel):
+    title: str = Field(min_length=3)
+    description: str = Field(min_length=3, max_length=100)
+    priority: int = Field(gt=0, lt=6)
+    complete: bool = Field(default=False)
+
+    # model_config = {
+    #     "json_schema_extra": {
+    #         "examples": {
+    #             "title": "example todo",
+    #             "description": "example description",
+    #             "priority": 3,
+    #             "complete": False,
+    #         }
+    #     }
+    # }
+
+
+# GET methods
+@router.get("/", status_code=status.HTTP_200_OK)
+async def read_all(db: db_dependency):
+    return db.query(Todos).all()
+
+
+@router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
+async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if todo_model is not None:
+        return todo_model
+    raise HTTPException(status_code=404, detail="Todo not found.")
+
+
+# POST methods
+@router.post("/todo", status_code=status.HTTP_201_CREATED)
+async def create_todo(db: db_dependency, todo_request: TodoRequest):
+    todo_model = Todos(**todo_request.model_dump())
+    db.add(todo_model)
+    db.commit()
+
+
+# PUT methods
+@router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def update_todo(
+        db: db_dependency,
+        todo_request: TodoRequest,
+        todo_id: int = Path(gt=0),
+):
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if todo_model is None:
+        raise HTTPException(status_code=404, detail="Todo not found.")
+
+    todo_model.title = todo_request.title
+    todo_model.description = todo_request.description
+    todo_model.priority = todo_request.priority
+    todo_model.complete = todo_request.complete
+
+    db.add(todo_model)
+    db.commit()
+
+
+# DELETE methods
+@router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+
+    if todo_model is None:
+        raise HTTPException(status_code=404, detail="Todo not found")
+
+    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.commit()
